Refactor enhanced -> admin_approval
