+++ test_app/settings_test.py
@@ -20,7 +20,7 @@ DEBUG = True
 ALLOWED_HOSTS = ['*']
 SECRET_KEY = '_'
 SITE_ID = 1
-ROOT_URLCONF = 'test_app.urls_enhanced'
+ROOT_URLCONF = 'test_app.urls_admin_approval'
 
 TEMPLATES = [
     {
